back_end/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
flask-restx==1.1.0
spacy==3.6.1
nltk==3.8.1
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
PyPDF2==3.0.3
fake-useragent==1.1.0
python-docx==0.8.11